Auth

+++ src/routes/deleteListing.js
@@ -1,4 +1,5 @@
 import { db } from "../database";
+import * as admin from 'firebase-admin';
 
 
 export const deleteListing={
@@ -6,9 +7,13 @@ export const deleteListing={
     path:'/api/listings/{id}',
     handler:async(req, h)=>{
         const {id} = req.params;
+        const token = req.headers.authtoken;
+        const user = await admin.auth().verifyIdToken(token);
+        const userId = user.user_id;
         await db.query(
-            'DELETE FROM `buy-and-sell`.listings WHERE id=?',
-            [id],
+            'DELETE FROM `buy-and-sell`.listings WHERE id=? AND user_id=?',
+            [id,userId],
+            
         );
 
         return {message:'Successfully deleted!'};
